Implementando repositório genérico para operações basicas
