+++ src/NossoCalendario.Domain/Base/IRepository.cs
@@ -8,6 +8,8 @@ namespace NossoCalendario.Domain.Base
     {
         IUnitOfWork UnitOfWork { get; }
         Task<TEntity> GetBy(Expression<Func<TEntity, bool>> predicate);
+        Task Insert(TEntity entity);
+        Task Update(TEntity entity);
     }
 
 }
