+++ src/NossoCalendario.Domain/Interfaces/IUsuarioRepository.cs
@@ -2,13 +2,12 @@
 using NossoCalendario.Domain.Entities;
 using System;
 using System.Linq.Expressions;
+using System.Threading.Tasks;
 
 namespace NossoCalendario.Domain.Interfaces
 {
     public interface IUsuarioRepository : IRepository<Usuario>
     {
-        void Insert(Usuario usuario);
-        void Update(Usuario usuario);
 
     }
 }
